+++ ramos/2_OBJETIVOS/kruskal.h
@@ -12,6 +12,6 @@ bool maiorIgualQuefloat(float a, float b);
 
 void intercala(float xl, float yl, float xll, float yll, int p, int q, int r, Aresta **v, int size, int direto);
 void mergesort(float xl, float yl, float xll, float yll, Aresta **v, int size, int direto);
-bool kruskal (Grafo *g, int  *A, float xl, float yl, float xll, float yll,float &custo,  float &x, float &y,int direto);
+bool kruskal (Grafo *g, int  *A, float xl, float yl, float xll, float yll,long double &custo,  float &x, float &y,int direto);
 
 #endif
